Fix shell syscall ABI, keyboard input, and tree formatting
add missing 64-bit signed alias for Galio kernel types
fix syscall declarations for pointer-sized arguments and execve ABI
correct pipe dup reference counting bug
validate execve argv/envp usage at the syscall boundary
restore guarded keyboard input path to avoid duplicate-key reads
allow ls and tree commands to accept optional path arguments
render directory tree with clear ASCII connectors using | and -
keep build working across shell and kernel changes

+++ include/syscall/syscall.h
@@ -68,7 +68,7 @@ u8 validate_user_buffer(const void *ptr, u32 length, u8 write);
 u8 validate_user_string(const char *str, u32 max_length);
 
 /* Extended syscall implementations */
-u32 syscall_execve(const char *path, char *const argv[], char *const envp[]);
+uintptr_t syscall_execve(const char *path, char *const argv[], char *const envp[]);
 i32 syscall_pipe(i32 pipefd[2]);
 u32 syscall_dup(u32 oldfd);
 i32 syscall_dup2(u32 oldfd, u32 newfd);
